requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
djangorestframework==3.10.0
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
python-http-client==3.1.0
sendgrid==6.0.5
django-cors-headers==3.0.2
html2text==2019.8.11